feat: error handeling for cart and authentication

+++ Kadoo/Cart/views.py
@@ -44,22 +44,27 @@ class AddPlantToCart(APIView):
  def post(self, request, format='json'):
   serializer = CartItemSerializer(data=request.data)
   if serializer.is_valid():
-   #Get All Data
-   userToAdd = request.user
-   PlantToAdd = Plant.objects.get(id=serializer.data["id"])
-   PlantToAddCount = serializer.data["count"]
-   PlantToAddDescription = serializer.data["description"]
-   #Check if it Exists
-   if PlantCartModel.objects.filter(user=userToAdd, plant_item=PlantToAdd, is_approved=False).exists():
-    return response.Response("This Plant is Alraedy On The Cart", status=status.HTTP_400_BAD_REQUEST)
-   #Make Cart Plant Item
-   PlantCartItem = PlantCartModel.objects.create(user=userToAdd, plant_item=PlantToAdd, plant_count=PlantToAddCount, description=PlantToAddDescription)
-   #Save Plant Item
-   PlantCartItem.save()
-   if PlantCartItem:
-    json = serializer.data
-    return response.Response(json, status=status.HTTP_201_CREATED)
-   return response.Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    #Get All Data
+    if request.user.is_anonymous:
+      return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
+    userToAdd = request.user
+    if Plant.objects.filter(id=serializer.data["id"]).exists() == False:
+      return response.Response("This plant does NOT Exist!", status=status.HTTP_422_UNPROCESSABLE_ENTITY)
+    PlantToAdd = Plant.objects.get(id=serializer.data["id"])
+    PlantToAddCount = serializer.data["count"]
+    PlantToAddDescription = serializer.data["description"]
+    #Check if it Exists
+    if PlantCartModel.objects.filter(user=userToAdd, plant_item=PlantToAdd, is_approved=False).exists():
+      return response.Response("This Plant is Alraedy On The Cart", status=status.HTTP_400_BAD_REQUEST)
+    #Make Cart Plant Item
+    PlantCartItem = PlantCartModel.objects.create(user=userToAdd, plant_item=PlantToAdd, plant_count=PlantToAddCount, description=PlantToAddDescription)
+    #Save Plant Item
+    PlantCartItem.save()
+    if PlantCartItem:
+      json = serializer.data
+      return response.Response(json, status=status.HTTP_201_CREATED)
+    return response.Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+  return response.Response("Data is Not Valid!", status=status.HTTP_400_BAD_REQUEST)
 
 #Create Tool
 class AddToolToCart(APIView):
@@ -67,7 +72,11 @@ class AddToolToCart(APIView):
   serializer = CartItemSerializer(data=request.data)
   if serializer.is_valid():
    #Get All Data
+   if request.user.is_anonymous:
+    return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
    userToAdd = request.user
+   if Tool.objects.filter(id=serializer.data["id"]).exists() == False:
+     return response.Response("This Tool does NOT Exist!", status=status.HTTP_422_UNPROCESSABLE_ENTITY)
    ToolToAdd = Tool.objects.get(id=serializer.data["id"])
    ToolToAddCount = serializer.data["count"]
    ToolToAddDescription = serializer.data["description"]
@@ -90,54 +99,80 @@ class AddToolToCart(APIView):
 #Read ALL Plants
 class GetAllPlantsCart(APIView):
  def get(self, request, format='json'):
+  if request.user.is_anonymous:
+    return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
   userToGetCart = request.user
   PlantsCartItems = PlantCartModel.objects.filter(user=userToGetCart)
+  if PlantsCartItems.count == 0:
+    return response.Response("No bought item nor in cart!", status=status.HTTP_200_OK);
   serializer = PlantCartSerializer(PlantsCartItems, many=True)
   return response.Response(serializer.data, status=status.HTTP_200_OK)
 
 #Read All Tools
 class GetAllToolsCart(APIView):
  def get(self, request, format='json'):
+  if request.user.is_anonymous:
+    return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
   userToGetCart = request.user
   ToolsCartItems = ToolCartModel.objects.filter(user=userToGetCart)
+  if ToolsCartItems.count == 0:
+    return response.Response("No bought item nor in cart!", status=status.HTTP_200_OK);
   serializer = ToolCartSerializer(ToolsCartItems, many=True)
   return response.Response(serializer.data, status=status.HTTP_200_OK)
 
 #Read ALL Unapproved Plants
 class GetUnapprovedPlantsCart(APIView):
  def get(self, request, format='json'):
+  if request.user.is_anonymous:
+    return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
   userToGetCart = request.user
   PlantsCartItems = PlantCartModel.objects.filter(user=userToGetCart, is_approved=False)
+  if PlantsCartItems.count == 0:
+    return response.Response("There is No Plant in Cart!", status=status.HTTP_404_NOT_FOUND);
   serializer = PlantCartSerializer(PlantsCartItems, many=True)
   return response.Response(serializer.data, status=status.HTTP_200_OK)
 
 #Read All Unapproved Tools
 class GetUnapprovedToolsCart(APIView):
  def get(self, request, format='json'):
+  if request.user.is_anonymous:
+    return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
   userToGetCart = request.user
   ToolsCartItems = ToolCartModel.objects.filter(user=userToGetCart, is_approved=False)
+  if ToolsCartItems.count == 0:
+    return response.Response("There is No Tool in Cart!", status=status.HTTP_404_NOT_FOUND);
   serializer = ToolCartSerializer(ToolsCartItems, many=True)
   return response.Response(serializer.data, status=status.HTTP_200_OK)
 
 #Read ALL Approved Plants
 class GetApprovedPlantsCart(APIView):
  def get(self, request, format='json'):
+  if request.user.is_anonymous:
+    return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
   userToGetCart = request.user
   PlantsCartItems = PlantCartModel.objects.filter(user=userToGetCart, is_approved=True)
+  if PlantsCartItems.count == 0:
+    return response.Response("No bought plant up to now!", status=status.HTTP_200_OK);
   serializer = PlantCartSerializer(PlantsCartItems, many=True)
   return response.Response(serializer.data, status=status.HTTP_200_OK)
 
 #Read All Approved Tools
 class GetApprovedToolsCart(APIView):
  def get(self, request, format='json'):
+  if request.user.is_anonymous:
+    return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
   userToGetCart = request.user
   ToolsCartItems = ToolCartModel.objects.filter(user=userToGetCart, is_approved=True)
+  if ToolsCartItems.count == 0:
+    return response.Response("No bought tool up to now!", status=status.HTTP_200_OK);
   serializer = ToolCartSerializer(ToolsCartItems, many=True)
   return response.Response(serializer.data, status=status.HTTP_200_OK)
 
 #Raed The Price
 class GetPriceCart(APIView):
  def get(self, request, format='json'):
+  if request.user.is_anonymous:
+    return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
   price = 0
   userToGetCart = request.user
   PlantsCartItems = PlantCartModel.objects.filter(user=userToGetCart, is_approved=False)
@@ -157,10 +192,12 @@ class GetPriceCart(APIView):
 #Raed The Count
 class GetCountCart(APIView):
  def get(self, request, format='json'):
+  if request.user.is_anonymous:
+    return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
   count = 0
   userToGetCart = request.user
-  count = count + PlantCartModel.objects.filter(user=userToGetCart, is_approved=False).count()
-  count = count + ToolCartModel.objects.filter(user=userToGetCart, is_approved=False).count()
+  count += PlantCartModel.objects.filter(user=userToGetCart, is_approved=False).count()
+  count += ToolCartModel.objects.filter(user=userToGetCart, is_approved=False).count()
   return response.Response(count, status=status.HTTP_200_OK)
 
 #------- UPDATE
@@ -172,11 +209,17 @@ class UpdatePlantToCart(APIView):
   serializer = CartItemSerializer(data=request.data)
   if serializer.is_valid():
    #Get All Data
+   if request.user.is_anonymous:
+    return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
    userToUpdate = request.user
+   if Plant.objects.filter(id=serializer.data["id"]).exists() == False:
+     return response.Response("This plant does NOT Exist!", status=status.HTTP_422_UNPROCESSABLE_ENTITY)
    PlantToUpdate = Plant.objects.get(id=serializer.data["id"])
    PlantToUpdateCount = serializer.data["count"]
    PlantToUpdateDescription = serializer.data["description"]
    #Make Cart Plant Item
+   if PlantCartModel.objects.filter(user=userToUpdate, plant_item=PlantToUpdate, is_approved=False).exists() == False:
+     return response.Response("This plant does NOT Exist in cart!", status=status.HTTP_404_NOT_FOUND)
    PlantCartItem = PlantCartModel.objects.get(user=userToUpdate, plant_item=PlantToUpdate, is_approved=False)
    PlantCartItem.plant_count = PlantToUpdateCount
    PlantCartItem.description = PlantToUpdateDescription
@@ -186,6 +229,7 @@ class UpdatePlantToCart(APIView):
     json = serializer.data
     return response.Response(json, status=status.HTTP_201_CREATED)
    return response.Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+  return response.Response("Data is Not Valid!", status=status.HTTP_400_BAD_REQUEST)
 
 #Update Tool
 class UpdateToolToCart(APIView):
@@ -193,11 +237,17 @@ class UpdateToolToCart(APIView):
   serializer = CartItemSerializer(data=request.data)
   if serializer.is_valid():
    #Get All Data
+   if request.user.is_anonymous:
+    return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
    userToUpdate = request.user
+   if Tool.objects.filter(id=serializer.data["id"]).exists() == False:
+     return response.Response("This Tool does NOT Exist!", status=status.HTTP_422_UNPROCESSABLE_ENTITY)
    ToolToUpdate = Tool.objects.get(id=serializer.data["id"])
    ToolToUpdateCount = serializer.data["count"]
    ToolToUpdateDescription = serializer.data["description"]
    #Make Cart Plant Item
+   if ToolCartModel.objects.filter(user=userToUpdate, tool_item=ToolToUpdate, is_approved=False).exists() == False:
+     return response.Response("This Tool does NOT Exist in cart!", status=status.HTTP_404_NOT_FOUND)
    ToolCartItem = ToolCartModel.objects.get(user=userToUpdate, tool_item=ToolToUpdate, is_approved=False)
    ToolCartItem.tool_count = ToolToUpdateCount
    ToolCartItem.description = ToolToUpdateDescription
@@ -207,6 +257,7 @@ class UpdateToolToCart(APIView):
     json = serializer.data
     return response.Response(json, status=status.HTTP_201_CREATED)
    return response.Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+  return response.Response("Data is Not Valid!", status=status.HTTP_400_BAD_REQUEST)
 
 #------- DELETE
 ###############
@@ -217,14 +268,19 @@ class RemovePlantFromCart(APIView):
   serializer = CartItemIdSerializer(data=request.data)
   if serializer.is_valid():
    #Get Plant Data
+   if request.user.is_anonymous:
+    return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
    UserToDeleteItem = request.user
+   if Plant.objects.filter(id=serializer.data["id"]).exists() == False:
+     return response.Response("This plant does NOT Exist!", status=status.HTTP_422_UNPROCESSABLE_ENTITY)
    PlantToRemove = Plant.objects.get(id=serializer.data["id"])
    #Make Cart Plant Item
    try:
     PlantCartModel.objects.filter(user=UserToDeleteItem, plant_item=PlantToRemove, is_approved=False).delete()
    except ProtectedError:
-    return response.Response("Can NOT Remove This Plant", status=status.HTTP_400_BAD_REQUEST)
-   return response.Response("Plant Removed",  status=status.HTTP_200_OK)
+    return response.Response("Can NOT Remove This Plant! this may does NOT Exist", status=status.HTTP_400_BAD_REQUEST)
+   return response.Response("Plant removed from the cart.",  status=status.HTTP_200_OK)
+  return response.Response("Data is Not Valid!", status=status.HTTP_400_BAD_REQUEST)
 
 #Delete Tool
 class RemoveToolFromCart(APIView):
@@ -232,18 +288,25 @@ class RemoveToolFromCart(APIView):
   serializer = CartItemIdSerializer(data=request.data)
   if serializer.is_valid():
    #Get Plant Data
+   if request.user.is_anonymous:
+    return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
    UserToDeleteItem = request.user
+   if Plant.objects.filter(id=serializer.data["id"]).exists() == False:
+     return response.Response("This plant does NOT Exist!", status=status.HTTP_422_UNPROCESSABLE_ENTITY)
    ToolToRemove = Tool.objects.get(id=serializer.data["id"])
    #Make Cart Plant Item
    try:
     ToolCartModel.objects.filter(user=UserToDeleteItem, tool_item=ToolToRemove, is_approved=False).delete()
    except ProtectedError:
     return response.Response("Can NOT Remove This Tool", status=status.HTTP_400_BAD_REQUEST)
-   return response.Response("Tool Removed",  status=status.HTTP_200_OK)
+   return response.Response("Tool removed from the cart.",  status=status.HTTP_200_OK)
+  return response.Response("Data is Not Valid!", status=status.HTTP_400_BAD_REQUEST)
 
 #Delte All Plants
 class RemoveAllPlantsFromCart(APIView):
  def delete(self, request, format='json'):
+   if request.user.is_anonymous:
+    return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
    UserToDeleteItem = request.user
    #Delete All
    try:
@@ -255,6 +318,8 @@ class RemoveAllPlantsFromCart(APIView):
 #Delete All Tools
 class RemoveAllToolsFromCart(APIView):
  def delete(self, request, format='json'):
+   if request.user.is_anonymous:
+    return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
    UserToDeleteItem = request.user
    #Delete All
    try:
@@ -268,9 +333,13 @@ class ApproveAllInCart(APIView):
  def post(self, request, format='json'):
    serializer = DesicriptionSerializer(data=request.data)
    if serializer.is_valid():
+    if request.user.is_anonymous:
+      return response.Response("Anonymous User: You should first login.", status=status.HTTP_401_UNAUTHORIZED)
     UserToApproveCart = request.user
     PlantsToApprove = PlantCartModel.objects.filter(user=UserToApproveCart, is_approved=False)
     ToolsToApprove = ToolCartModel.objects.filter(user=UserToApproveCart, is_approved=False)
+    if PlantsToApprove.count == 0 and ToolsToApprove.count == 0:
+     return response.Response("Cart is Empthy!", status=status.HTTP_400_BAD_REQUEST)
     OrderDescription = serializer.data["description"]
     price = 0
     for PlantItem in PlantsToApprove:
